frontend/requirements.txt
streamlit==1.48.1
requests==2.32.5
pillow==11.3.0
streamlit-webrtc
av
aiortc
opencv-python-headless

sievedata
mediapipe opencv-python
